@@ -3,22 +3,21 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@react-spring/three": "^9.4.0",
-    "@react-three/cannon": "^4.4.1",
-    "@react-three/drei": "^8.3.1",
+    "@react-three/drei": "^8.18.10",
     "@react-three/fiber": "^7.0.24",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
-    "framer-motion": "^5.5.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "framer-motion": "^6.2.8",
+    "framer-motion-3d": "^6.2.8",
+    "react": "^18.0.0-rc.3",
+    "react-dom": "^18.0.0-rc.3",
     "react-helmet-async": "^1.2.3",
     "react-responsive": "^9.0.0-beta.5",
     "react-router-dom": "^5.3.0",
     "react-scripts": "5.0.0",
-    "three": "^0.135.0",
-    "three-stdlib": "^2.6.4",
+    "three": "^0.137.0",
+    "three-stdlib": "^2.8.8",
     "web-vitals": "^2.1.2"
   },
   "scripts": {
