@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import * as ReactDOMClient from "react-dom/client";
 import "./index.css";
 import App from "./App";
 
@@ -12,7 +12,7 @@ const browserHistory = createBrowserHistory({
   basename: ".",
 });
 
-ReactDOM.render(
+ReactDOMClient.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
     <HelmetProvider>
       <Router history={browserHistory}>
@@ -24,6 +24,5 @@ ReactDOM.render(
         />
       </Router>
     </HelmetProvider>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
