@@ -1,6 +1,6 @@
 import { useScroll } from "@react-three/drei";
 import { useFrame } from "@react-three/fiber";
-import { LayoutCamera } from "framer-motion/three";
+import { LayoutCamera } from "framer-motion-3d";
 import { useState } from "react";
 
 const Camera = ({ isPhone }) => {
