@@ -1,4 +1,4 @@
-import { motion } from "framer-motion/three";
+import { motion } from "framer-motion-3d";
 
 const Lights = () => {
   return (
