@@ -1,11 +1,7 @@
-import { MotionCanvas } from "framer-motion/three";
-import {
-  Preload,
-  ScrollControls,
-  useCursor,
-} from "@react-three/drei";
+import { MotionCanvas } from "framer-motion-3d";
+import { Preload, ScrollControls, useCursor } from "@react-three/drei";
 import Lights from "./Lights";
-import { useState } from "react";
+import { Suspense, useState } from "react";
 import PostsPresenter from "./Posts/PostsPresenter";
 
 const PostsScene = ({ isPhone }) => {
@@ -20,13 +16,20 @@ const PostsScene = ({ isPhone }) => {
       dpr={[1, 2]}
       style={{
         backgroundColor: theme ? "#F1F4F8" : "#0E0B07",
+        width: "100%",
       }}
     >
-      <Preload all />
-      <ScrollControls pages={isPhone ? 2.55 : 1.55} damping={4} distance={2}>
-        <PostsPresenter setCursor={setCursor} theme={theme} isPhone={isPhone} />
-      </ScrollControls>
-      <Lights />
+      <Suspense fallback={null}>
+        <Preload all />
+        <ScrollControls pages={isPhone ? 2.55 : 1.55} damping={4} distance={2}>
+          <PostsPresenter
+            setCursor={setCursor}
+            theme={theme}
+            isPhone={isPhone}
+          />
+        </ScrollControls>
+        <Lights />
+      </Suspense>
     </MotionCanvas>
   );
 };
